--HG--
branch : akd

+++ scripts/pymod-process.py
@@ -1541,7 +1541,7 @@ def generate_getset_code(tp):
 	getter = """
 local PyObject * %(getname)s(PyObject *obj, void *v)
 {
-	int offset = (int)v;
+	int offset = (long)v;
 	%(ctype)s *ptr = (%(ctype)s*)(((char*)obj)+offset);
 """ % vars()
 	try:
@@ -1558,7 +1558,7 @@ local PyObject * %(getname)s(PyObject *obj, void *v)
 	setter = """
 local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 {
-	int offset = (int)v;
+	int offset = (long)v;
 	%(ctype)s *ptr = (%(ctype)s*)(((char*)obj)+offset);
 """ % vars()
 
@@ -1566,7 +1566,7 @@ local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 		setter = """
 local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 {
-	int offset = (int)v;
+	int offset = (long)v;
 	int *ptr = (int*)(((char*)obj)+offset);
 """ % vars()
